+++ src/commands/settings_commands/settings.js
@@ -160,74 +160,36 @@ const getSettings = function getSettings (data)
 
    };
 
-
-   // -----------
-   // Update bot
-   // -----------
-
-   const updateBot = function updateBot (data)
-   {
-
-      // Const activeGuilds = data.client.guilds.array();
-      // Data.color = "info";
-      // Data.text = `Updating bot for **${activeGuilds.length}** servers.`;
-      // Return sendMessage(data);
-      //
-      // ActiveGuilds.forEach(guild =>
-      // {
-      //   Guild.owner.send(
-      //   "Hello, this bot has been updated to a new version.\n " +
-      //   "More info: https://ritabot.gg/whats-new/#new-in-121\n");
-      // });
-      data.message.delete({"timeout": time.short}).catch((err) => console.log(
-         "UpdateBot Command Message Deleted Error, command.send.js = ",
-         err
-      ));
-      return data.message.channel.send({"embed": {
-         "author": {
-            "icon_url": data.client.user.displayAvatarURL(),
-            "name": data.client.user.username
-         },
-         "color": 13107200,
-         "description": ":no_entry_sign: This command has been disabled"
-
-      }}).then((msg) =>
-      {
-
-         msg.delete({"timeout": time.long}).catch((err) => console.log(
-            "UpdateBot Bot Message Deleted Error, settings.js = ",
-            err
-         ));
-
-      });
-
-   };
-
    // -----------------
    // DM server owners
    // -----------------
-   // Announcements not possible until D.js v12
 
    /*
-   Const announcement = async function(data)
+   // eslint-disable-next-line no-unused-vars
+   const Announcement = async function Announcement (data)
    {
-      const guildArray = Array.from(bot.client.guilds.values());
-      var i;
+
+      const guildArray = Array.from(data.client.guilds.cache());
+      let i;
       for (i = 0; i < guildArray.length; i += 1)
       {
+
          console.log("Hello");
          const guild = await guildArray[i];
-         var owner = await guild.ownerID;
-         // eslint-disable-next-line quotes
-         owner = Number(owner)
-         // eslint-disable-next-line no-undef
+         let owner = await guild.ownerID;
+         owner = Number(owner);
          owner = owner.replace(/([0-9]+)/g, "$1");
          console.log("Done");
-         await data.client.users.get(owner).send("Testing").catch((err) =>
-         {
-            console.log(err);
-         });
+         await data.client.users.get(owner).send("Testing").
+            catch((err) =>
+            {
+
+               console.log(err);
+
+            });
+
       }
+
    };
    */
 
@@ -257,11 +219,11 @@ const getSettings = function getSettings (data)
    // --------------------------
 
    const validSettings = {
-   // "announcement": announcement,
+      // "announce": announcement,
       "listservers": listServers,
       "setlang": setLang,
-      "updatebot": updateBot,
       "updatedb": updateDB
+
    };
 
    const settingParam = data.cmd.params.split(" ")[0].toLowerCase();