+++ src/commands/utility_commands/help.js
@@ -11,6 +11,7 @@ const time = {
    "long": 60000,
    "short": 5000
 };
+const helpFooter = `**For more help join our official [support server](<https://ritabot.gg>)**\n\n`;
 
 // -------------
 // Help Section
@@ -20,7 +21,7 @@ const helpSection = function helpSection (data)
 {
 
    const section =
-      `${data.icon}  **[${data.title}](${data.link})**\n\n`;
+      `${data.icon}  **[${data.title}](${data.link})**\n\`\`\`${data.value}\`\`\`\n`;
 
    return section;
 
@@ -30,7 +31,7 @@ const helpSection = function helpSection (data)
 // Help Text
 // ----------
 
-const helpMessage = function helpMessage (config, botname, param)
+const helpMessage = function helpMessage (config, param)
 {
 
    // ---------
@@ -41,12 +42,10 @@ const helpMessage = function helpMessage (config, botname, param)
    const long = config.translateCmd;
 
    const info =
-   `**${botname} Bot - v.${config.version}**\n` +
-   `Translates Discord messages (based on \`Google API\`).\n\n` +
    "```md\n" +
-   `# All the commands can be found by entering the following:\n` +
-   `* ${cmd} help commands for a list of all avaliable commands or \n` +
-   `* ${cmd} help modules for a list of avaliable command modules` +
+   `# For a full list of commands, enter the following:\n` +
+   `* ${cmd} help commands\n` +
+   `* ${cmd} help modules` +
    "```\n\n";
 
    // ------------
@@ -55,120 +54,53 @@ const helpMessage = function helpMessage (config, botname, param)
 
    const basics =
    helpSection({
-      "args": null,
-      "cmd": null,
       config,
-      "example": "!t help react for examples",
-      "help": "react",
       "icon": ":flag_white:",
-      "link": "<https://ritabot.gg/trans-reac/>",
-      "title": "Translate by Reacting"
+      "link": `<https://ritabot.gg>`,
+      "title": "Translate by Reacting",
+      "value": `${cmd} help react`
    }) +
    helpSection({
-      "args": "to [lang] from [lang]: [msg]",
-      "cmd": "this",
       config,
-      "example": "to french from english: hello",
-      "help": "custom",
       "icon": ":abc:",
-      "link": "<https://ritabot.gg/trans-cust/>",
-      "title": "Translate Custom Text"
+      "link": `<https://ritabot.gg>`,
+      "title": "Translate Custom Text",
+      "value": `${cmd} help react`
    }) +
    helpSection({
-      "args": "[count] from [lang] to [lang]",
-      "cmd": "last",
       config,
-      "example": "3 from german to spanish",
-      "help": "last",
       "icon": ":arrow_double_up:",
-      "link": "<https://ritabot.gg/trans-last/>",
-      "title": "Translate Last Message"
+      "link": `<https://ritabot.gg>`,
+      "title": "Translate Last Message",
+      "value": `${cmd} help react`
    }) +
    helpSection({
-      "args": "from [lang] to [lang] for [@/#]",
-      "cmd": "channel",
       config,
-      "example": "from hebrew to arabic for me",
-      "help": "auto",
       "icon": ":hash:",
-      "link": "<https://ritabot.gg/trans-auto/>",
-      "title": "Translate Channel (Automatic)"
+      "link": `<https://ritabot.gg>`,
+      "title": "Translate Channel (Automatic)",
+      "value": `${cmd} help auto`
    }) +
    helpSection({
-      "args": "stats [server/global]",
-      "cmd": "stats",
       config,
-      "example": "",
-      "help": "stats",
       "icon": ":bar_chart:",
-      "link": "<https://ritabot.gg/trans-misc/#statistics>",
-      "title": "Stats"
+      "link": `<https://ritabot.gg>`,
+      "title": "Stats",
+      "value": `${cmd} help stats`
    }) +
    helpSection({
-      "args": "setLang to [lang]",
-      "cmd": "settings",
       config,
-      "example": "setLang to italian",
-      "help": "settings",
       "icon": ":gear:",
-      "link": "<https://ritabot.gg/trans-sett/>",
-      "title": "Settings"
+      "link": `<https://ritabot.gg>`,
+      "title": "Settings",
+      "value": `${cmd} help settings`
    }) +
    helpSection({
-      "args": "",
-      "cmd": "misc",
       config,
-      "example": "",
-      "help": "misc",
       "icon": ":robot:",
-      "link": "<https://ritabot.gg/trans-misc/>",
-      "title": "Misc. Settings"
-   }) +
-   helpSection({
-      "args": "",
-      "cmd": "Tasks",
-      config,
-      "example": "",
-      "help": "Tasks",
-      "icon": ":clipboard:",
-      "link": "<https://ritabot.gg/trans-misc/#translations>",
-      "title": "Tasks"
-   }) +
-   helpSection({
-      "args": "",
-      "cmd": "readme",
-      config,
-      "example": "",
-      "help": "readme",
-      "icon": ":bookmark_tabs:",
-      "link": "<https://ritabot.gg/secure/>",
-      "title": "ReadMe"
-   }) +
-   helpSection({
-      config,
-      "icon": "🙋🏽‍♀️",
-      "link": "<https://github.com/RitaBot-Project/RitaBot/issues>",
-      "title": "Report Bugs / Request Features"
-   }) +
-   helpSection({
-      "args": "oc",
-      "cmd": "donate",
-      config,
-      "example": "",
-      "help": "donate",
-      "icon": ":dollar: ",
-      "link": "<https://opencollective.com/ritabot-project>",
-      "title": "Donate Via Open Collective"
-   }) +
-   helpSection({
-      "args": "github",
-      "cmd": "donate",
-      config,
-      "example": "",
-      "help": "donate",
-      "icon": ":dollar: ",
-      "link": "<https://github.com/sponsors/RitaBot-Project>",
-      "title": "Donate Via Github"
+      "link": `<https://ritabot.gg>`,
+      "title": "Misc. Settings",
+      "value": `${cmd} help misc`
    });
 
    // ----------------
@@ -177,32 +109,35 @@ const helpMessage = function helpMessage (config, botname, param)
 
    const modules =
    `__**Modules**__\n\n` +
-   `These will help you look through ${botname}'s commands.` +
-   "```md\n" +
 
+   "```md\n" +
    `# Translation Help Modules\n` +
    `* ${cmd} help auto\n` +
    `* ${cmd} help react\n` +
    `* ${cmd} help last\n` +
    `* ${cmd} help this\n\n` +
-
+   "```" +
+   "```md\n" +
    `# Translation Style Modules\n` +
    `* ${cmd} help embed\n` +
    `* ${cmd} help bot2bot\n\n` +
-
+   "```" +
+   "```md\n" +
    `# Information Modules\n` +
    `* ${cmd} help stats\n` +
    `* ${cmd} help tasks\n` +
    `* ${cmd} help settings (Dev Only)\n\n` +
-
+   "```" +
+   "```md\n" +
    `# Other Modules\n` +
    `* ${cmd} help misc\n` +
    `* ${cmd} help report\n` +
    `* ${cmd} help readme\n` +
    `* ${cmd} help donate\n` +
    `* ${cmd} help debug\n` +
+   `* ${cmd} help dev\n` +
    `* ${cmd} help prefix\n` +
-   "```";
+   "```\n";
 
    // -----------------
    // ReadMe + Discord
@@ -228,7 +163,7 @@ const helpMessage = function helpMessage (config, botname, param)
    `You can report issues in Github here:\n` +
    `- https://github.com/RitaBot-Project/RitaBot/issues\n\n` +
    `Rita also has a support server with it's developers and contributors: \n` +
-   `- https://discord.gg/mgNR64R \n` +
+   `- [Offical Support Server](https://discord.gg/mgNR64R) \n` +
    "\n\n";
 
    // --------------
@@ -237,7 +172,7 @@ const helpMessage = function helpMessage (config, botname, param)
 
    const commands =
    `__**All Available Commands**__\n\n` +
-   `This is a list of all avaliable commands\n` +
+   `This is a list of all available commands\n` +
    "```md\n" +
    `# Translation Commands\n` +
    `* ${cmd} this to [lang] from [lang]: [msg]\n` +
@@ -245,7 +180,8 @@ const helpMessage = function helpMessage (config, botname, param)
    `* ${cmd} channel to [lang] from [lang] for [me/@/#]\n` +
    `* ${cmd} auto to [lang] for [me/@/#]\n` +
    `* ${cmd} stop for [me/@/#]\n\n` +
-
+   "```" +
+   "```md\n" +
    `# Information Commands\n` +
    `* ${cmd} tasks\n` +
    `* ${cmd} stats [server/global]\n` +
@@ -253,24 +189,32 @@ const helpMessage = function helpMessage (config, botname, param)
    `* ${cmd} proc\n` +
    `* ${cmd} shards\n` +
    `* ${cmd} id\n` +
-   `* ${cmd} list\n\n` +
-
+   `* ${cmd} list\n` +
+   "```" +
+   "```md\n" +
    `# Style Commands\n` +
    `* ${cmd} embed [on/off]\n` +
    `* ${cmd} bot2bot [on/off]\n` +
    `* ${cmd} prefix [prefix]\n\n` +
-
+   "```" +
+   "```md\n" +
    `# Misc Commands\n` +
    `* ${cmd} invite\n` +
-   `* ${cmd} help modules\n\n` +
-
+   `* ${cmd} help modules\n` +
+   `* ${cmd} check [ServerID]\n` +
+   `* ${cmd} perm [user/bot]\n` +
+   `* ${cmd} create [channelName]\n` +
+   `* ${cmd} history \n\n` +
+   "```" +
+   "```md\n" +
    `# Debug Commands\n` +
-   `* ${cmd} debug [on/pff]\n` +
+   `* ${cmd} debug [on/off]\n` +
    `* ${cmd} stats debug (Admin Only)\n\n` +
-
+   "```" +
+   "```md\n" +
    `# Want to Support RITA\n` +
    `* ${cmd} donate [oc/github]\n\n` +
-   "```";
+   "```\n";
 
    // ----------------------
    // Custom message (this)
@@ -279,23 +223,25 @@ const helpMessage = function helpMessage (config, botname, param)
    const custom =
    `__**Translate Custom Message**__\n\n` +
    `Translates a custom message entered by user.\n` +
-   "```md\n" +
 
+   "```md\n" +
    `# Command\n` +
    `* ${cmd} this: [msg] \n` +
    `* ${cmd} this to [lang] from [lang]: [msg] \n\n` +
-
+   "```" +
+   "```md\n" +
    `# Parameters\n` +
    `* to [lang] - defaults to server default language\n` +
    `* to [lang, lang, ...] - translates to multiple languages\n` +
    `* from [lang] - defaults to automatic detection\n\n` +
-
+   "```" +
+   "```md\n" +
    `# Examples\n` +
    `* ${cmd} this: bonjour \n` +
    `* ${cmd} this to spanish: hello world \n` +
    `* ${cmd} this to arabic, hebrew: I love you \n` +
    `* ${cmd} this to de from en: how are you? \n` +
-   "```";
+   "```\n";
 
    // --------------------
    // Flag Emoji Reaction
@@ -303,13 +249,11 @@ const helpMessage = function helpMessage (config, botname, param)
 
    const react =
    `__**Translate by reaction**__\n\n` +
-   `# Add a message reaction with the flag of the language` +
-   `you wish to translate to\n\n` +
-   `# Examples\n` +
-   `:flag_gb: = English\n` +
-   `:flag_fr: = French\n\n` +
-   ` Please note that only official country flags are registered`;
 
+   "```md\n" +
+   `# Add a message reaction with the flag of the language you wish to translate to\n\n` +
+   ` Please note that only official country flags are registered\n` +
+   "```\n";
    // --------------------
    // Last Message (last)
    // --------------------
@@ -317,26 +261,31 @@ const helpMessage = function helpMessage (config, botname, param)
    const last =
    `__**Translate Last Message(s)**__\n\n` +
    `Translates last message chain(s) in channel. A chain is a collection of ` +
-   `messages by the same author, to keep things simple.\n` +
-   "```md\n" +
+   `messages by the same author, to keep things simple.\n\n` +
 
+   "```md\n" +
+   `# Due to a technical fault, this command is disabled until further notice. Sorry for any inconvenience this may cause.` +
+   "```\n" +
+   "```md\n" +
    `# Command\n` +
    `* ${cmd} last \n` +
    `* ${cmd} last [n] to [lang] from [lang] \n\n` +
-
+   "```" +
+   "```md\n" +
    `# Parameters\n` +
    `* to [lang] - defaults to server default language\n` +
    `* to [lang, lang, ...] - translates to multiple languages\n` +
    `* from [lang] - defaults to automatic detection\n` +
    `* [n] - number of chains to translate, default is 1\n` +
    `* [-n] - negative number means only one chain is translated\n\n` +
-
+   "```" +
+   "```md\n" +
    `# Examples\n` +
    `* ${cmd} last 2 \n` +
    `* ${cmd} last to english \n` +
    `* ${cmd} last to english, german, french \n` +
    `* ${cmd} last -6 to english from german` +
-   "```";
+   "```\n";
 
    // -------------------------
    // Auto translate (channel)
@@ -347,24 +296,26 @@ const helpMessage = function helpMessage (config, botname, param)
    `Automatically translates any new messages in channel and forwards them ` +
    `to you. Admins/mods can set forwarding to other channels or users in ` +
    `server. Messages in forwarded channels will also be sent back to origin*.` +
-   "```md\n" +
 
+   "```md\n" +
    `# Command\n` +
    `* ${cmd} channel \n` +
    `* ${cmd} channel to [lang] from [lang] for [me/@/#] \n` +
    `* ${cmd} stop for [me/@/#] \n\n` +
-
+   "```" +
+   "```md\n" +
    `# Parameters\n` +
    `* to [lang] - defaults to server default language\n` +
    `* from [lang] -  language to translate from, now includes 'auto'\n` +
    `* for [me/@/#] - defaults to "me", admins can use mentions \n\n` +
-
+   "```" +
+   "```md\n" +
    `# Examples\n` +
-   `* ${cmd} channel to english from chinese \n` +
-   `* ${cmd} channel to en from de for #englishChannel \n` +
-   `* ${cmd} channel to de from en for @steve \n` +
-   `* ${cmd} channel to en from ru for #ch1, #ch2, #usr1 \n` +
-   "```";
+   `* ${cmd} channel from english to chinese \n` +
+   `* ${cmd} channel from en to de for #englishChannel \n` +
+   `* ${cmd} channel from de to fr for @steve \n` +
+   `* ${cmd} channel from en to es for #ch1, #ch2, #usr1 \n` +
+   "```\n";
 
    // --------------
    // Tasks Command
@@ -373,15 +324,16 @@ const helpMessage = function helpMessage (config, botname, param)
    const tasks =
    `__**Channel Tasks**__\n\n` +
    `Displays translation tasks of the current channel` +
-   "```md\n" +
 
+   "```md\n" +
    `# Displays translation tasks of the current channel\n` +
    `* ${cmd} tasks\n\n` +
-
+   "```" +
+   "```md\n" +
    `# Displays translation tasks of specified channel\n` +
    `* COMING IN FUTURE UPDATE \n` +
    `* ${cmd} tasks for [#channel]\n` +
-   "```";
+   "```\n";
 
    // ----------------------
    // Auto translate (stop)
@@ -391,22 +343,24 @@ const helpMessage = function helpMessage (config, botname, param)
    `__**Stop Auto Translation**__\n\n` +
    `Terminates auto-translation of channel for you. ` +
    `Admins/mods can stop for other channels or users in server.` +
-   "```md\n" +
 
+   "```md\n" +
    `# Command\n` +
    `* ${cmd} stop \n` +
    `* ${cmd} stop for [me/@/#/all] \n\n` +
-
+   "```" +
+   "```md\n" +
    `# Parameters\n` +
    `* for [me/@/#/all] - defaults to "me" \n\n` +
-
+   "```" +
+   "```md\n" +
    `# Examples\n` +
    `* ${cmd} stop \n` +
    `* ${cmd} stop for me \n` +
    `* ${cmd} stop for @usr1 \n` +
    `* ${cmd} stop for #ch1 \n` +
    `* ${cmd} stop for all \n` +
-   "```";
+   "```\n";
 
    // -------------
    // Misc Command
@@ -414,27 +368,40 @@ const helpMessage = function helpMessage (config, botname, param)
 
    const misc =
    `__**Miscellaneous Commands**__\n\n` +
-   "```md\n" +
 
+   "```md\n" +
    `# Help\n` +
    `* ${cmd} help\n` +
    `* ${cmd} help [command]\n\n` +
-
+   "```" +
+   "```md\n" +
    `# Links\n` +
    `* ${cmd} invite\n\n` +
-
+   "```" +
+   "```md\n" +
    `# Supported Languages\n` +
    `* ${cmd} list\n\n` +
-
+   "```" +
+   "```md\n" +
+   `# Permission Check\n` +
+   `* ${cmd} perm [user/bot]\n\n` +
+   "```" +
+   "```md\n" +
    `# Donate\n` +
    `* ${cmd} donate [oc/github]\n\n` +
-
+   "```" +
+   "```md\n" +
    `# Prefix\n` +
    `* ${cmd} prefix\n\n` +
-
-   `# Prefix\n` +
+   "```" +
+   "```md\n" +
+   `# ID Info\n` +
    `* ${cmd} id\n\n` +
-   "```";
+   "```" +
+   "```md\n" +
+   `# Blacklist Check\n` +
+   `* ${cmd} check [ServerID]\n\n` +
+   "```\n";
 
    // -----------------
    // Settings Command
@@ -442,19 +409,11 @@ const helpMessage = function helpMessage (config, botname, param)
 
    const settings =
    `__**Settings**__\n\n` +
+
    "```md\n" +
-   `# Defunct Command\n` +
-   `* ${cmd} settings updatebot\n\n` +
-
-   `# Update bot Database with new columns\n` +
-   `* ${cmd} settings updatedb\n\n` +
-
    `# Set default server language\n` +
    `* ${cmd} settings setLang to [lang]\n\n` +
-
-   `# Displays list of servers the bot is in\n` +
-   `* ${cmd} settings listservers\n\n` +
-   "```";
+   "```\n";
 
    // -------------------
    // Statistics Command
@@ -462,17 +421,23 @@ const helpMessage = function helpMessage (config, botname, param)
 
    const stats =
    `__**Statistics**__\n\n` +
-   "```md\n" +
 
+   "```md\n" +
    `# Statistics\n` +
    `* ${cmd} version \n` +
    `* ${cmd} stats \n` +
    `* ${cmd} stats global \n` +
    `* ${cmd} stats server \n\n` +
-
+   "```" +
+   "```md\n" +
    `* Admin Only - Do not use in public channels\n` +
-   `* ${cmd} stats debug \n` +
-   "```";
+   `* ${cmd} stats server [ServerID]\n` +
+   `* ${cmd} stats debug \n\n` +
+   "```" +
+   "```md\n" +
+   `# Parameters\n` +
+   `* [ServerID] - Raw Server ID\n\n` +
+   "```\n";
 
    // -----------
    // Embed Help
@@ -480,19 +445,21 @@ const helpMessage = function helpMessage (config, botname, param)
 
    const embed =
    `__**Message Embed Styles**__\n\n` +
-   "```md\n" +
 
+   "```md\n" +
    `# Command\n` +
    `* ${cmd} embed [on/off]\n\n` +
-
+   "```" +
+   "```md\n" +
    `# Parameters\n` +
    `* on - Turns on Embed Translation\n` +
    `* off - Turns on Webhook Translation Sending\n\n` +
-
+   "```" +
+   "```md\n" +
    `# Examples\n` +
    `* ${cmd} embed on \n` +
    `* ${cmd} embed off \n` +
-   "```";
+   "```\n";
 
    // -------------
    // Bot2bot Help
@@ -500,20 +467,22 @@ const helpMessage = function helpMessage (config, botname, param)
 
    const bot2bot =
    `__**Bot to Bot Translation**__\n\n` +
+
    "```md\n" +
    `* Usually 90% of bots ignore other bot messages but this feature attemptes to translate them.\n\n` +
-
    `# Command\n` +
    `* ${cmd} bot2bot [on/off]\n\n` +
-
+   "```" +
+   "```md\n" +
    `# Parameters\n` +
    `* on - Turns on Bot2Bot Translations\n` +
    `* off - Turns off Bot2Bot Translations\n\n` +
-
+   "```" +
+   "```md\n" +
    `# Examples\n` +
    `* ${cmd} bot2bot on \n` +
    `* ${cmd} bot2bot off \n` +
-   "```";
+   "```\n";
 
    // -----------------
    // Debug Command
@@ -521,19 +490,21 @@ const helpMessage = function helpMessage (config, botname, param)
 
    const debug =
    `__**Debug Webhook**__\n\n` +
-   "```md\n" +
 
+   "```md\n" +
    `# Command\n` +
    `* ${cmd} debug [on/off]\n\n` +
-
+   "```" +
+   "```md\n" +
    `# Parameters\n` +
    `* on - Turns debug webhook on\n` +
    `* off - Turns debug webhook off\n\n` +
-
+   "```" +
+   "```md\n" +
    `# Examples\n` +
    `* ${cmd} debug on \n` +
    `* ${cmd} debug off \n` +
-   "```";
+   "```\n";
 
    // ---------------
    // Prefix Command
@@ -541,22 +512,24 @@ const helpMessage = function helpMessage (config, botname, param)
 
    const prefix =
    `__**Prefix**__\n\n` +
-   "```md\n" +
 
+   "```md\n" +
    `# Command\n` +
    `* ${cmd} prefix [prefix]\n` +
    `* ${long} prefix [prefix]\n\n` +
-
+   "```" +
+   "```md\n" +
    `# Parameters\n` +
    `* [prefix] - Changes the prefix of your bot commands\n` +
    `* reset - Reset your prefix back to default\n\n` +
-
+   "```" +
+   "```md\n" +
    `# Examples\n` +
    `* ${cmd} prefix $tr - your prefix would now be $tr \n` +
    `* ${long} prefix $tr - your prefix would now be $tr \n` +
    `* ${cmd} prefix reset - resets your prefix back to !tr\n` +
    `* ${long} prefix reset - resets your prefix back to !tr\n` +
-   "```";
+   "```\n";
 
    // ----------
    // Donations
@@ -564,8 +537,8 @@ const helpMessage = function helpMessage (config, botname, param)
 
    const donate =
    `__**Want to Donate to RITA's Development **__\n\n` +
-   "```md\n" +
 
+   "```md\n" +
    `# Donate\n` +
    `* Becoming a Sponsor, Supporter or Backer of RitaBot will \n` +
    `* allow us to continue development long into the future, \n` +
@@ -577,7 +550,7 @@ const helpMessage = function helpMessage (config, botname, param)
    `* You can Donate at Open Collective\n` +
    `* ${cmd} donate oc\n\n` +
    `* Thank you for your continued support - RITA Dev Team` +
-   "```";
+   "```\n";
 
    // ---------------
    // Create Command
@@ -585,19 +558,46 @@ const helpMessage = function helpMessage (config, botname, param)
 
    const create =
    `__**Create**__\n\n` +
-   "```md\n" +
 
+   "```md\n" +
    `# Command\n` +
    `* ${cmd} create [channelName]\n` +
    `* ${long} create [channelName]\n\n` +
-
+   "```" +
+   "```md\n" +
    `# Parameters\n` +
-   `* [channelName] - sets name of new channel\n` +
-
+   `* [channelName] - sets name of new channel\n\n` +
+   "```" +
+   "```md\n" +
    `# Examples\n` +
    `* ${cmd} create bob - your new channel would now be bob \n` +
    `* ${long} create bob - your new channel would now be bob \n` +
-   "```";
+   "```\n";
+   // ---------------
+   // Create Command
+   // ---------------
+
+   const dev =
+   `__**Developers Only**__\n\n` +
+   "```md\n" +
+   `# Commands\n` +
+   `* ${cmd} eject [ServerID]\n` +
+   `* ${cmd} blacklist [ServerID]\n` +
+   `* ${cmd} unblacklist [ServerID]\n` +
+   `* ${cmd} server [ServerID]\n` +
+   `* ${cmd} settings updatedb\n` +
+   `* ${cmd} settings listservers\n\n` +
+   "```" +
+   "```md\n" +
+   `# Parameters\n` +
+   `* [ServerID] - Raw Server ID\n\n` +
+   "```" +
+   "```md\n" +
+   `# Examples\n` +
+   `* ${cmd} eject [ServerID] - Eject Rita from [ServerID]\n` +
+   `* ${cmd} blacklist [ServerID] - [ServerID] Will be blacklisted\n\n` +
+   "```\n";
+
 
    // ----------------
    // Proccess result
@@ -612,6 +612,7 @@ const helpMessage = function helpMessage (config, botname, param)
       create,
       custom,
       debug,
+      dev,
       donate,
       embed,
       last,
@@ -657,7 +658,8 @@ module.exports = function run (data)
 
    let getHelpWith = "basics";
 
-   if (data.cmd.params === "settings")
+   // eslint-disable-next-line no-constant-condition
+   if (data.cmd.params === "dev")
    {
 
       const cleanParam = data.cmd.params.toLocaleLowerCase().trim();
@@ -666,13 +668,6 @@ module.exports = function run (data)
       if (!process.env.DISCORD_BOT_OWNER_ID.includes(data.message.author.id))
       {
 
-         getHelpWith = cleanParam;
-         data.text = helpMessage(
-            data.config,
-            data.bot.username,
-            getHelpWith
-         );
-
          console.log("Insufficient Permission");
          data.message.delete({"timeout": time.short}).catch((err) => console.log(
             "Command Message Deleted Error, help.js = ",
@@ -680,10 +675,6 @@ module.exports = function run (data)
          ));
          richEmbedMessage.
             setColor(colors.get(data.color)).
-            setAuthor(
-               data.bot.username,
-               data.bot.displayAvatarURL()
-            ).
             setDescription("This command is available only to Developers. \n\n").
             setTimestamp().
             setFooter("This message will self-destruct in one minute");
@@ -706,9 +697,8 @@ module.exports = function run (data)
       getHelpWith = cleanParam;
       data.text = helpMessage(
          data.config,
-         data.bot.username,
          getHelpWith
-      );
+      ) + helpFooter;
 
       // -------------
       // Send message
@@ -722,22 +712,20 @@ module.exports = function run (data)
 
       data.text = helpMessage(
          data.config,
-         data.bot.username,
          getHelpWith
-      );
+      ) + helpFooter;
       return sendMessage(data);
 
    }
-   else if (data.cmd.params !== "settings" || null)
+   else if (data.cmd.params !== null)
    {
 
       const cleanParam = data.cmd.params.toLocaleLowerCase().trim();
       getHelpWith = cleanParam;
       data.text = helpMessage(
          data.config,
-         data.bot.username,
          getHelpWith
-      );
+      ) + helpFooter;
 
       // -------------
       // Send message
@@ -749,9 +737,8 @@ module.exports = function run (data)
 
    data.text = helpMessage(
       data.config,
-      data.bot.username,
       getHelpWith
-   );
+   ) + helpFooter;
 
    // -------------
    // Send message