+++ src/core/logger.js
@@ -92,6 +92,13 @@ const errorLog = function errorLog (error, subtype, id)
 
    }
 
+   if (errorTitle === "unhandled")
+   {
+
+      return console.log(`DEBUG: Error ${errorTitle} Suppressed`);
+
+   }
+
    hookSend({
       "color": "err",
       // eslint-disable-next-line no-useless-concat