+++ src/core/send.js
@@ -13,7 +13,7 @@ const fn = require("./helpers");
 const db = require("./db");
 const logger = require("./logger");
 const discord = require("discord.js");
-const webHookName = "Translator Messaging System";
+const webHookName = "RITA Messaging System";
 const error = require("./error");
 
 // -----------------
@@ -134,6 +134,7 @@ const checkPerms = function checkPerms (data, sendBox)
                   Server: **${data.channel.guild.name}** \n
                   Channel: **${forwardChannel.name}**\n
                   Chan ID: **${forwardChannel.id}**\n
+                  Server ID: **${data.channel.guild.id}**\n
                   Owner: **${data.channel.guild.owner}**\n
                   The server owner has been notified . \n`
             });
@@ -388,9 +389,14 @@ const embedOn = function embedOn (data)
                   Server: **${data.guild.name}** \n
                   Channel: **${data.channel.name}**\n
                   Chan ID: **${data.channel.id}**\n
+                  Server ID: **${data.channel.guild.id}**\n
                   Owner: **${data.channel.guild.owner}**\n
                   The server owner has been notified. \n`
-                  });
+                  }).catch((err) => console.log(
+                     "error",
+                     err,
+                     "warning"
+                  ));
 
                }
 
@@ -405,7 +411,11 @@ const embedOn = function embedOn (data)
                   Channel: **${data.channel.name}**\n
                   Chan ID: **${data.channel.id}**\n
                   Owner: **${data.channel.guild.owner}**\n`
-                  });
+                  }).catch((err) => console.log(
+                     "error",
+                     err,
+                     "warning"
+                  ));
 
                }
 
@@ -420,7 +430,11 @@ const embedOn = function embedOn (data)
                   Channel: **${data.channel.name}**\n
                   Chan ID: **${data.channel.id}**\n
                   Owner: **${data.channel.guild.owner}**\n`
-                  });
+                  }).catch((err) => console.log(
+                     "error",
+                     err,
+                     "warning"
+                  ));
 
                }
 
@@ -450,13 +464,22 @@ const embedOn = function embedOn (data)
                            "```prolog\nServer > Privacy Settings > " +
                            "'Allow direct messages from server members'\n```");
 
-                  });
+                  }).catch((err) => console.log(
+                     "error",
+                     err,
+                     "warning"
+                  ));
 
                }
 
                logger("error", errMsg, "warning", data.message.channel.guild.name);
 
-            });
+            }).
+            catch((err) => console.log(
+               "error",
+               err,
+               "warning"
+            ));
 
       }
       else if (data.attachments.array().length > 0)
@@ -545,7 +568,7 @@ const embedOff = function embedOff (data)
             "avatarURL": data.message.author.displayAvatarURL(),
             files,
             "username": data.message.author.username
-         });
+         }).catch((err) => console.log("error", err, "send", data.message.guild.name));
 
       }