+++ src/message.js
@@ -91,7 +91,7 @@ module.exports = function run (config, message)
 
       }
 
-      console.log(`--m.js--- Empty Message Error: ----1----\nServer: ${message.channel.guild.name},\nChannel: ${message.channel.id} - ${message.channel.name},\nMessage ID: ${message.id},\nContent: ${message.content},\nWas Image: ${message.attachments},\nwas Embed: ${message.embeds},\nSender: ${message.member.displayName} - ${message.member.id},\nTimestamp: ${message.createdAt}\n----------------------------------------`);
+      // console.log(`--m.js--- Empty Message Error: ----1----\nServer: ${message.channel.guild.name},\nChannel: ${message.channel.id} - ${message.channel.name},\nMessage ID: ${message.id},\nContent: ${message.content},\nWas Image: ${message.attachments},\nWas Embed: ${message.embeds},\nSender: ${message.member.displayName} - ${message.member.id},\nTimestamp: ${message.createdAt}\n----------------------------------------`);
 
    }
 
@@ -122,13 +122,16 @@ module.exports = function run (config, message)
          message.content = `${message.content}\n${message.embeds[0].description}`;
 
       }
+
+      /*
       else if (message.content === "" || message.content === " ")
       {
 
-         console.log(`--m.js--- Empty Message Error: ----2----\nServer: ${message.channel.guild.name},\nChannel: ${message.channel.id} - ${message.channel.name},\nMessage ID: ${message.id},\nContent: ${message.content},\nWas Image: ${message.attachments},\nwas Embed: ${message.embeds},\nSender: ${message.member.displayName} - ${message.member.id},\nTimestamp: ${message.createdAt}\n----------------------------------------`);
+         console.log(`--m.js--- Empty Message Error: ----2----\nServer: ${message.channel.guild.name},\nChannel: ${message.channel.id} - ${message.channel.name},\nMessage ID: ${message.id},\nContent: ${message.content},\nWas Image: ${message.attachments},\nWas Embed: ${message.embeds},\nSender: ${message.member.displayName} - ${message.member.id},\nTimestamp: ${message.createdAt}\n----------------------------------------`);
          return;
 
       }
+      */
 
    }
 
@@ -185,6 +188,27 @@ module.exports = function run (config, message)
       }
 
    }
+
+   // ---------------------
+   // Blacklist Redundancy
+   // ---------------------
+   const serverID = data.message.guild.id;
+
+   db.getServerInfo(
+      serverID,
+      function getServerInfo (server)
+      {
+
+         if (server[0].blacklisted === true)
+         {
+
+            data.message.guild.leave();
+            console.log(`Blacklist Redundancy, Server ${serverID} ejected`);
+
+         }
+
+      }
+   );
 
    // ------------------
    // Proccess Commands